requirements.txt
Django==1.11
django-contrib-comments==1.8.0
django-crispy-forms==1.6.1
django-filter==1.0.4
djangorestframework==3.6.3
Markdown==2.6.8
mtgsdk==1.3.0
olefile==0.44
Pillow==4.2.1
pytz==2017.2
